+++ dev-packages/application-manager/src/generator/webpack-generator.ts
@@ -195,26 +195,6 @@ module.exports = {
             {
                 test: /\\.plist$/,
                 type: 'asset/resource'
-            },
-            {
-                test: /\\.js$/,
-                // include only es6 dependencies to transpile them to es5 classes
-                include: /vscode-ws-jsonrpc|vscode-jsonrpc|vscode-languageserver-protocol|vscode-languageserver-types/,
-                use: {
-                    loader: 'babel-loader',
-                    options: {
-                        presets: ['@babel/preset-env'],
-                        plugins: [
-                            // reuse runtime babel lib instead of generating it in each js file
-                            '@babel/plugin-transform-runtime',
-                            // ensure that classes are transpiled
-                            '@babel/plugin-transform-classes'
-                        ],
-                        // see https://github.com/babel/babel/issues/8900#issuecomment-431240426
-                        sourceType: 'unambiguous',
-                        cacheDirectory: true
-                    }
-                }
             }
         ]
     },