+++ packages/plugin-ext/src/main/browser/view/tree-view-widget.tsx
@@ -281,7 +281,7 @@ export class TreeViewWidget extends TreeViewWelcomeWidget {
         });
 
         const children = this.getCaption(node);
-        return React.createElement('div', attrs, ...children);
+        return React.createElement('div', attrs, children);
     }
 
     protected getCaption(node: TreeNode): React.ReactNode {