+++ packages/plugin-ext/src/plugin/types-impl.ts
@@ -24,7 +24,7 @@
 
 import { UUID } from '@theia/core/shared/@phosphor/coreutils';
 import { illegalArgument } from '../common/errors';
-import * as theia from '@theia/plugin';
+import type * as theia from '@theia/plugin';
 import { URI as CodeURI, UriComponents } from '@theia/core/shared/vscode-uri';
 import { relative } from '../common/paths-util';
 import { startsWithIgnoreCase } from '@theia/core/lib/common/strings';
@@ -33,6 +33,7 @@ import { SymbolKind } from '../common/plugin-api-rpc-model';
 import { FileSystemProviderErrorCode, markAsFileSystemProviderError } from '@theia/filesystem/lib/common/files';
 import * as paths from 'path';
 import { ObjectsTransferrer } from '../common/rpc-protocol';
+import { es5ClassCompat } from '../common/types';
 
 /**
  * A reviver that takes URI's transferred via JSON.stringify() and makes
@@ -52,6 +53,7 @@ export function reviver(key: string | undefined, value: any): any {
  * We cannot simply upgrade the dependency, because the curent version 3.x
  * is not compatible with our current codebase
  */
+@es5ClassCompat
 export class URI extends CodeURI implements theia.Uri {
     protected constructor(scheme: string, authority?: string, path?: string, query?: string, fragment?: string, _strict?: boolean);
     protected constructor(components: UriComponents);
@@ -114,6 +116,7 @@ export class URI extends CodeURI implements theia.Uri {
     }
 }
 
+@es5ClassCompat
 export class Disposable {
     private disposable: undefined | (() => void);
 
@@ -207,6 +210,7 @@ export enum SourceControlInputBoxValidationType {
     Information = 2
 }
 
+@es5ClassCompat
 export class ColorTheme implements theia.ColorTheme {
     constructor(public readonly kind: ColorThemeKind) { }
 }
@@ -233,6 +237,7 @@ export namespace TextEditorSelectionChangeKind {
     }
 }
 
+@es5ClassCompat
 export class Position {
     private _line: number;
     private _character: number;
@@ -394,6 +399,7 @@ export class Position {
     }
 }
 
+@es5ClassCompat
 export class Range {
     protected _start: Position;
     protected _end: Position;
@@ -528,6 +534,7 @@ export class Range {
     }
 }
 
+@es5ClassCompat
 export class Selection extends Range {
     private _anchor: Position;
     private _active: Position;
@@ -579,6 +586,7 @@ export enum EnvironmentVariableMutatorType {
     Prepend = 3
 }
 
+@es5ClassCompat
 export class SnippetString {
 
     static isSnippetString(thing: {}): thing is SnippetString {
@@ -660,11 +668,13 @@ export class SnippetString {
     }
 }
 
+@es5ClassCompat
 export class ThemeColor {
     constructor(public id: string) {
     }
 }
 
+@es5ClassCompat
 export class ThemeIcon {
 
     static readonly File: ThemeIcon = new ThemeIcon('file');
@@ -723,6 +733,7 @@ export enum ConfigurationTarget {
     Memory
 }
 
+@es5ClassCompat
 export class RelativePattern {
 
     base: string;
@@ -753,6 +764,7 @@ export enum IndentAction {
     Outdent = 3
 }
 
+@es5ClassCompat
 export class TextEdit {
 
     protected _range: Range;
@@ -861,6 +873,7 @@ export enum CompletionItemKind {
     TypeParameter = 24
 }
 
+@es5ClassCompat
 export class CompletionItem implements theia.CompletionItem {
 
     label: string;
@@ -883,6 +896,7 @@ export class CompletionItem implements theia.CompletionItem {
     }
 }
 
+@es5ClassCompat
 export class CompletionList {
 
     isIncomplete?: boolean;
@@ -907,16 +921,7 @@ export enum DebugConsoleMode {
     MergeWithParent = 1
 }
 
-export class DiagnosticRelatedInformation {
-    location: Location;
-    message: string;
-
-    constructor(location: Location, message: string) {
-        this.location = location;
-        this.message = message;
-    }
-}
-
+@es5ClassCompat
 export class Location {
     uri: URI;
     range: Range;
@@ -942,6 +947,17 @@ export class Location {
     }
 }
 
+@es5ClassCompat
+export class DiagnosticRelatedInformation {
+    location: Location;
+    message: string;
+
+    constructor(location: Location, message: string) {
+        this.location = location;
+        this.message = message;
+    }
+}
+
 export enum DiagnosticTag {
     Unnecessary = 1,
 }
@@ -950,6 +966,7 @@ export enum CompletionItemTag {
     Deprecated = 1,
 }
 
+@es5ClassCompat
 export class Diagnostic {
     range: Range;
     message: string;
@@ -977,6 +994,7 @@ export enum MarkerTag {
     Unnecessary = 1,
 }
 
+@es5ClassCompat
 export class ParameterInformation {
     label: string | [number, number];
     documentation?: string | MarkdownString;
@@ -987,6 +1005,7 @@ export class ParameterInformation {
     }
 }
 
+@es5ClassCompat
 export class SignatureInformation {
     label: string;
     documentation?: string | MarkdownString;
@@ -1005,6 +1024,7 @@ export enum SignatureHelpTriggerKind {
     ContentChange = 3,
 }
 
+@es5ClassCompat
 export class SignatureHelp {
     signatures: SignatureInformation[];
     activeSignature: number;
@@ -1015,6 +1035,7 @@ export class SignatureHelp {
     }
 }
 
+@es5ClassCompat
 export class Hover {
 
     public contents: MarkdownString[] | theia.MarkedString[];
@@ -1044,6 +1065,7 @@ export enum DocumentHighlightKind {
     Write = 2
 }
 
+@es5ClassCompat
 export class DocumentHighlight {
 
     public range: Range;
@@ -1060,6 +1082,7 @@ export class DocumentHighlight {
 
 export type Definition = Location | Location[];
 
+@es5ClassCompat
 export class DocumentLink {
 
     range: Range;
@@ -1080,6 +1103,7 @@ export class DocumentLink {
     }
 }
 
+@es5ClassCompat
 export class CodeLens {
 
     range: Range;
@@ -1119,6 +1143,7 @@ export enum CodeActionTriggerKind {
     Automatic = 2,
 }
 
+@es5ClassCompat
 export class CodeActionKind {
     private static readonly sep = '.';
 
@@ -1155,6 +1180,7 @@ export enum TextDocumentSaveReason {
     FocusOut = 3
 }
 
+@es5ClassCompat
 export class CodeAction {
     title: string;
 
@@ -1207,6 +1233,7 @@ export interface FileTextEdit {
     metadata?: WorkspaceEditMetadata;
 }
 
+@es5ClassCompat
 export class WorkspaceEdit implements theia.WorkspaceEdit {
 
     private _edits = new Array<FileOperation | FileTextEdit | undefined>();
@@ -1319,6 +1346,7 @@ export class WorkspaceEdit implements theia.WorkspaceEdit {
     }
 }
 
+@es5ClassCompat
 export class TreeItem {
 
     label?: string | theia.TreeItemLabel;
@@ -1356,6 +1384,7 @@ export enum SymbolTag {
     Deprecated = 1
 }
 
+@es5ClassCompat
 export class SymbolInformation {
 
     static validate(candidate: SymbolInformation): void {
@@ -1400,6 +1429,7 @@ export class SymbolInformation {
     }
 }
 
+@es5ClassCompat
 export class DocumentSymbol {
 
     static validate(candidate: DocumentSymbol): void {
@@ -1439,6 +1469,7 @@ export enum CommentThreadCollapsibleState {
     Expanded = 1
 }
 
+@es5ClassCompat
 export class QuickInputButtons {
     static readonly Back: theia.QuickInputButton = {
         iconPath: {
@@ -1450,6 +1481,7 @@ export class QuickInputButtons {
     };
 }
 
+@es5ClassCompat
 export class FileDecoration {
 
     static validate(d: FileDecoration): void {
@@ -1487,6 +1519,7 @@ export enum FileChangeType {
     Deleted = 3,
 }
 
+@es5ClassCompat
 export class FileSystemError extends Error {
 
     static FileExists(messageOrUri?: string | URI): FileSystemError {
@@ -1548,6 +1581,7 @@ export interface FileStat {
     readonly size: number;
 }
 
+@es5ClassCompat
 export class ProgressOptions {
     /**
      * The location at which progress should show.
@@ -1569,6 +1603,8 @@ export class ProgressOptions {
         this.location = location;
     }
 }
+
+@es5ClassCompat
 export class Progress<T> {
     /**
      * Report a progress update.
@@ -1594,6 +1630,7 @@ export enum ProgressLocation {
     Notification = 15
 }
 
+@es5ClassCompat
 export class ProcessExecution {
     private executionProcess: string;
     private arguments: string[];
@@ -1673,6 +1710,7 @@ export enum TaskRevealKind {
     Never = 3
 }
 
+@es5ClassCompat
 export class ShellExecution {
     private shellCommandLine: string;
     private shellCommand: string | theia.ShellQuotedString;
@@ -1746,6 +1784,7 @@ export class ShellExecution {
     }
 }
 
+@es5ClassCompat
 export class CustomExecution {
     private _callback: (resolvedDefintion: theia.TaskDefinition) => Thenable<theia.Pseudoterminal>;
     constructor(callback: (resolvedDefintion: theia.TaskDefinition) => Thenable<theia.Pseudoterminal>) {
@@ -1766,6 +1805,7 @@ export class CustomExecution {
     }
 }
 
+@es5ClassCompat
 export class TaskGroup {
     private groupId: string;
 
@@ -1809,6 +1849,7 @@ export enum TaskScope {
     Workspace = 2
 }
 
+@es5ClassCompat
 export class Task {
     private taskDefinition: theia.TaskDefinition;
     private taskScope: theia.TaskScope.Global | theia.TaskScope.Workspace | theia.WorkspaceFolder | undefined;
@@ -1994,10 +2035,12 @@ export class Task {
     }
 }
 
+@es5ClassCompat
 export class Task2 extends Task {
     detail?: string;
 }
 
+@es5ClassCompat
 export class DebugAdapterExecutable {
     /**
      * The command or path of the debug adapter executable.
@@ -2034,6 +2077,7 @@ export class DebugAdapterExecutable {
 /**
  * Represents a debug adapter running as a socket based server.
  */
+@es5ClassCompat
 export class DebugAdapterServer {
 
     /**
@@ -2068,6 +2112,7 @@ export enum LogLevel {
 /**
  * The base class of all breakpoint types.
  */
+@es5ClassCompat
 export class Breakpoint {
     /**
      * Is breakpoint enabled.
@@ -2109,6 +2154,7 @@ export class Breakpoint {
 /**
  * A breakpoint specified by a source location.
  */
+@es5ClassCompat
 export class SourceBreakpoint extends Breakpoint {
     /**
      * The source and line position of this breakpoint.
@@ -2127,6 +2173,7 @@ export class SourceBreakpoint extends Breakpoint {
 /**
  * A breakpoint specified by a function name.
  */
+@es5ClassCompat
 export class FunctionBreakpoint extends Breakpoint {
     /**
      * The name of the function to which this breakpoint is attached.
@@ -2142,6 +2189,7 @@ export class FunctionBreakpoint extends Breakpoint {
     }
 }
 
+@es5ClassCompat
 export class Color {
     readonly red: number;
     readonly green: number;
@@ -2156,6 +2204,7 @@ export class Color {
     }
 }
 
+@es5ClassCompat
 export class ColorInformation {
     range: Range;
     color: Color;
@@ -2172,6 +2221,7 @@ export class ColorInformation {
     }
 }
 
+@es5ClassCompat
 export class ColorPresentation {
     label: string;
     textEdit?: TextEdit;
@@ -2191,6 +2241,7 @@ export enum ColorFormat {
     HSL = 2
 }
 
+@es5ClassCompat
 export class FoldingRange {
     start: number;
     end: number;
@@ -2209,6 +2260,7 @@ export enum FoldingRangeKind {
     Region = 3
 }
 
+@es5ClassCompat
 export class SelectionRange {
 
     range: Range;
@@ -2257,6 +2309,7 @@ export enum UIKind {
     Web = 2
 }
 
+@es5ClassCompat
 export class CallHierarchyItem {
     _sessionId?: string;
     _itemId?: string;
@@ -2292,28 +2345,31 @@ export class CallHierarchyItem {
     }
 }
 
+@es5ClassCompat
 export class CallHierarchyIncomingCall {
 
     from: theia.CallHierarchyItem;
     fromRanges: theia.Range[];
 
-    constructor(item: theia.CallHierarchyItem, fromRanges: theia.Range[]) {
+    constructor(item: CallHierarchyItem, fromRanges: Range[]) {
         this.fromRanges = fromRanges;
         this.from = item;
     }
 }
 
+@es5ClassCompat
 export class CallHierarchyOutgoingCall {
 
     to: theia.CallHierarchyItem;
     fromRanges: theia.Range[];
 
-    constructor(item: theia.CallHierarchyItem, fromRanges: theia.Range[]) {
+    constructor(item: CallHierarchyItem, fromRanges: Range[]) {
         this.fromRanges = fromRanges;
         this.to = item;
     }
 }
 
+@es5ClassCompat
 export class TimelineItem {
     timestamp: number;
     label: string;
@@ -2331,6 +2387,7 @@ export class TimelineItem {
 
 // #region Semantic Coloring
 
+@es5ClassCompat
 export class SemanticTokensLegend {
     public readonly tokenTypes: string[];
     public readonly tokenModifiers: string[];
@@ -2345,6 +2402,7 @@ function isStrArrayOrUndefined(arg: any): arg is string[] | undefined {
     return ((typeof arg === 'undefined') || (Array.isArray(arg) && arg.every(e => typeof e === 'string')));
 }
 
+@es5ClassCompat
 export class SemanticTokensBuilder {
 
     private _prevLine: number;
@@ -2356,7 +2414,7 @@ export class SemanticTokensBuilder {
     private _tokenModifierStrToInt: Map<string, number>;
     private _hasLegend: boolean;
 
-    constructor(legend?: theia.SemanticTokensLegend) {
+    constructor(legend?: SemanticTokensLegend) {
         this._prevLine = 0;
         this._prevChar = 0;
         this._dataIsSortedAndDeltaEncoded = true;
@@ -2524,6 +2582,7 @@ export class SemanticTokensBuilder {
     }
 }
 
+@es5ClassCompat
 export class SemanticTokens {
     readonly resultId?: string;
     readonly data: Uint32Array;
@@ -2534,6 +2593,7 @@ export class SemanticTokens {
     }
 }
 
+@es5ClassCompat
 export class SemanticTokensEdit {
     readonly start: number;
     readonly deleteCount: number;
@@ -2546,6 +2606,7 @@ export class SemanticTokensEdit {
     }
 }
 
+@es5ClassCompat
 export class SemanticTokensEdits {
     readonly resultId?: string;
     readonly edits: SemanticTokensEdit[];